requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn==0.27.0

# Database & Migration
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.3.6

# AI & Processing
google-generativeai==0.3.2
pydantic-settings==2.1.0
PyMuPDF==1.23.21

# Crawler
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.1